Add clean-webpack-plugin and terser-webpack-plugin for optimizing the production build
Updated several files to increase the efficiency and cleanliness of the project's production build. In webpack.prod.js, terser-webpack-plugin is added to the optimization.minimizer array for code minification, and clean-webpack-plugin is introduced to clean the /dist folder before each build. This prevents the distribution of unnecessary or outdated files and helps to reduce the total bundle size.

The tsconfig.json file has been updated to modify the ECMAScript target and module code generation to ES5 and ES2015 respectively, a requirement for certain older browsers compatibility, and the package.json and pnpm-lock.yaml files have been updated to include clean-webpack-plugin and terser-webpack-plugin dependencies.

Moreover, in pnpm-lock.yaml, added a few other packages like '/@types/glob@7.2.0' and '/del@4.1.1', which are the dependencies of the newly added clean-webpack-plugin.