Remove terser-webpack-plugin from dependencies
The terser-webpack-plugin was removed from 'pnpm-lock.yaml' and 'package.json' files. The reason behind this removal is an update in our webpack.prod.js minimizer configuration. This change was implemented to optimize the identification of existing minimizers, thus improving the construction of the production build and maintaining a small bundle size. This enhances code readability and maintainability.